dev,plugins: Fix import types from .proto files
Gorums does not handle types defined in another .proto file. As the
fully qualified type name (includes the package name) is used, we
encounter a problem when a type is embedded in a struct. This is not a
problem if the type and Gorums service are in the same package, as the
package name is stripped. However when importing type definitions from
another package, invalid code is generated.

If all message types are external, the io and strings packages are never
imported, so these need to be included if no messages are locally
defined.

This fix makes these changes:

- typeName method changed to only return the type name. The package name
  is not part of the name.

- fqTypeName method retains old typeName functionality.

- RespName and ReqName changed to FQRespName and FQReqName. The values
  are the same.

- RespName is now the type name only, and is used where the output type
  is embedded in a struct.

- io and strings imports are only ignored if file contains any messages.

- As the .tmpl files have changed, plugins/gorums/templates.go have been
regenerated.