reorganized the config templates also
this could be viewed as a little risky, but i've fixed on bug
and the tests are passing, so i think it is a reasonable first
step towards splitting things on per method basis.

thus far, i've kept the same 'Future' and 'Correctable' etc
suffixes and the shared struct when a method is defined with
multiple options.