dev,plugins,testdata: Make per_node_arg take a base request
This makes the per_node_arg option work with the qf_with_req option. Each
invocation of the node function receives a copy of the base request. This
potentially makes it easier to work with larger messages where only a few fields
need to change per node.

make goldenanddev is run to reflect the changes in the generated files.