byzq: use custom_return_type and qf_with_req; general clean up
removed wts field not needed by AuthDataQ.Write() since we are
already matching replies with gRPC. Required fixing some tests
and the WriteQF. We now take the 'req' as argument to the QF
and check that a quorum of the servers replied with a correct
timestamp. (not sure if this is really needed?)

We also use the custom return type Content to avoid exposing
the signature fields of the reply Value.