Added NewResponseType shared for both exported and internal types.
Note: We want to retain the use of UnexportedTypeName across both
types used in the calltype_datatypes.tmpl and other calltypes
definitions, hence we store the same name in both TypeName and
UnexportedTypeName fields. This is ok, since we keep two slices,
one for exported and one for internal types.