added nil check to Node funcs; added Port() function.
the nil check prevents nil-related panics when a Node object
hasn't been initialized, i.e. is nil.

the Port() function helps to print debug statements that are
shorter, especially on localhost/127.0.0.1. this could also
be leveraged for multi-machine configurations as well, if
the different machines in a configuration use different port
numbers, encoding it such that it is uniquely determined
which machine it is based on the port.