cleanup: remove several custom template functions; use generic one
This commit replaces the use of custom template functions for
ensuring that imports are included in the various generated files.
That is, we now use {{use "time.Until" .GenFile}} to ensure that
the time package is imported. We do not need to use this function
for all identifiers of the imported packages, only one time per
package.