clean up makefile
First stab at cleaning up the makefile; still a few todo items
pending. Mainly made a few of the make targets depend on src files
so that they won't rebuild if src files didn't change; this is
especially relevant for the bundle target which is slow.

The other targets are relatively quick, and also not trivial to
make depend on src/dst files; hence, dev target is still phony.
The dev target is also the default (appears first in makefile).