moved future API methods to datatypes file.
This commit simplifies the generation of API methods associated
with a promise type, specifically FutureResponse and FutureMyResponse
in this case. I moved the API methods to be generated alongside
the actual struct type for the different Future types.
This avoids the need for the hasAPIType function, and makes things
slightly easier to maintain and understand (hopefully).

A similar change should be done for the correctable types.