moved devTypes to gengorums package and cleaned up special cases
This commit fixes the less than desirable use of special cases
in the generation of different dev files, such as node, qspec and
datatypes for quorum calls. Now these cases are handled from a
single source, the devTypes map. The use of this map may be
expanded to also include templates for the different quorum calls,
instead of have a separate gorumsCallTypeTemplates map.
But I'm doing this refactoring in two steps to avoid mistakes.