simplified gen; removing another map and added output stability
This commit fixes output stability, ensuring that multiple runs
of the same input proto file will produce the same output
(done by sorting the order in which input option types are generated)
Added a make target to check stability. This is done using the
zorums.proto (now copied also to tests/zorums folder).

Also eliminated the gorumsTypes map, since we acheive the same
with gorumsCallTypesInfo map, which also holds the non-option
types, such as node, qspec, and types, which only need a template.