first draft of rewrite with lots of debug prints
The plan is to clean this up and squash the commits when
merging into the rewrite branch.

This commit essentially tries to generalize the concept of
tagged calltypes, such as quorumcall with ordered option.

It's a bit tricky since currently, the big table of calltypes
must explicitly exclude other calltypes. A revision should
hopefully avoid this, or do a precheck to ensure that all
table entries has a mutually exclusive chkFn().