unified gorums file guard
This commit simplifies and unifies the guard for gorums proto
files, meaning that we now also check for Gorums reserved idents
in the proto message types also for dev files. Moreover, we
simplified the checking for Gorums method types in the proto
file; now there is only one method to do this instead of two.