removed qf_with_req gorums option
This commit removes the need to use the qf_with_req option.
That is, users must now implement their quorum functions with
the request object passed in always; they may not use it,
in which case they should ignore it by declaring their quorum
function with _. This approach seems to eliminate any performance
penalty from the extra argument that may not always be needed.