fixed a bug in async selection and upgraded grpc lib
This commit now generates code based on quorumcall, ordered and
async gorums options.
Unfortunately, this also lead me down a path to upgrade the lib.
Ended up forking the grpc lib, removing the module files in the
cmd/protoc-go-gen-grpc, because it didn't work as expected...
Hope this doesn't cause any problems; if it does, we can roll
back the change, and reapply only the more relevant changes.