make 'async_handler' the default
This also removes the need for the 'concurrent' option, as the same
behavoir can be achieved by starting a goroutine from the new handler.

The 'async_handler' pattern allows the user to decide whether the
handlers should run fully synchronously, or asynchronously. It also
preserves message ordering, i.e. handlers are called in the order that
the messages were received in, but do not necessarily return in the same
order.