Added a test for TLS
The server-side metadata allows us to check the auth type of the client.
This commit adds a test to see if TLS is working with Gorums.
Also, turns out that the reason the old TLS test broke was because the
included cert had expired. To avoid this problem in the future, the test
now generates its own certificate.