Refactored static code for into gorums package
This moves as much as possible of the static code included with the
generated code into the gorums package.
I have also refactored the quorum call such that the implementation is
in the gorums package too.
What remains in the generated code is only wrappers that convert between
application and generic types.