nodeStream: prevent deadlock when server stops receiving
If the server stops calling RecvMsg(), the SendMsg() call will
eventually start blocking. If this happens, the sendQ channels will
eventually fill up and block the QC.

To prevent this, we can use the QC context to abort the stream if
SendMsg() is blocking. The stream will then try to reconnect.