NodeStream: avoid CloseSend concurrently with SendMsg
The gRPC documentation states that it is unsafe to call CloseSend
concurrently with SendMsg. Instead, we can use context cancellation to
stop the stream. This requires making a new context each time we start
the stream.

In addition, I realized that orderedNodeStream.sendMsg() can run the
gRPC SendMsg() directly instead of in a goroutine, with the logic to
cancel the stream running in a goroutine instead. This seems to improve
performance.