Fix deadlock in orderedNodeStream
If a request context is cancelled in sendMsg, then the goroutine
monitoring sendMsg will cancel the stream. However, the channel 'c' that
communicates with the monitoring goroutine is unbuffered, which means that
sendMsg will deadlock when attempting to send a message on that channel.
The solution is to make the channel buffered.