Fixed bugs in PerNode transform func and added tests
This commit add tests for MulticastPerNode message transformation
functions. The commit also fixes a deadlock bug that prevented it
from unblocking the call when the message transformation returns
nil for a node, indicating that the node should not receive the
message. This also fixes a bug that would not actully use the
transformed message, and a nil check on the transformed message
was no longer possible with the new reflection-based protobuf library.