Made call types methods on gorums.Configuration
This makes quorum call types into methods on gorums.Configuration,
which simply is an embedding of []*Node. This makes call types
sligthly nicer to implement and there is less data to add to the
QuorumCallData struct. The next commit will correspondingly change
the single node call types to be methods on gorums.Node.