Added newCall+newReply methods on gorums.Configuration
This makes it possible to implement quorum call types using just the
gorums.Configuration type to reach the newCall and newReply methods
of the receiveQueue. The receiveQueue is global for a single manager
running on a client, but this receiveQueue instance is passed to all
nodes via the orderedNodeStream type. Hence, all nodes has access to
the same receiveQueue instance, and therefore in Configuration.newCall
we simply access this receiveQueue instance via the Configuration's
first node.

In conclusion, this makes it nicer to implement quourm call types
with access only to the gorums.Configuration type, and we avoid the
need to pass the Manager via the QuorumCallData struct.