NodeList and NodeMap are now passed to Configuration
This is a major redesign of how we pass node lists and node maps
to the system. This now happens only through the creation of a
new configuration, but still using the WithNodeList and WithNodeMap
options. These can no longer be passed to the manager. The manager
still keeps the pool of all nodes in all configurations. A benefit
from this change is that the manager can no longer fail with an error.

As this is a relatively major change, and there were fair amount
of usage examples that had to be updated, this also required a bit
of associated clean up.

When this rewrite was more or less finished I realized that we could
also provide WithNodeIDs, which could have avoided many of the
tricky rewrites. I plan to add WithNodeIDs in a follow up commit.

Fixes #30.