Added NodeListOption to RemoveNodes and Remove configs
This adds support for creating a new configuration based on an existing
configuration, removing nodes via the Configuration.RemoveNodes() method.
Also, we can create a new configuration based on two existing configuration
via the Configuration.Remove(Configuration) method, meaning that the new
configuration removes the nodes from the configuration given to the
method.