Added --help to protoc-gen-gorums and print to stdout
This cleans up the printing to stderr, and instead prints to stdout,
following recent changes in protoc-gen-go. Further, it adds --help
to direct users to the web page, if wondering why they installed the
protoc-gen-gorums command. This also removes some debug printing
when generating _gorums.pb.go files; now it is silent, similar to
protoc-gen-go. It does still print Generating bundle file though,
and diff results, if the template_static.go changed.

Note that in general we should not print info to stdout during code
generation; if debug output is needed, that should be sent to stderr.
However, for the scenarios we use stdout here, we always do os.Exit(),
in which case it is fine (and seemingly preferred).