add a ConfigurationFromRaw function
This function allows us to create a "clone" of a configuration with a
different QuorumSpec from the original.

It could also be used to clone a configuration that was created by a
different gorums service altogether.
 That is, imagine we have two .proto files describing separate services.
 If we create a configuration using the manager from service1,
 we can access the configuration's RawConfiguration field and use that
with ConfigurationFromRaw from service2.
Now both configuration objects will refer to the same nodes, but will
expose different methods.