Fixed bug in error handling
Previously there was no way to know whether or not a goroutine had
in fact been started; henceforth, the errChan would be closed by
the defer since the createReplicas() call would reach the return
before all goroutines had the chance to send their err, causing a
panic. In this new version, we wait for all goroutines to start
before we check for errors on the channel.