feat!: add support for errors.Is() on QuorumCallError
This refactors the QuorumCallError to support the errors.Is() func.
This is a breaking change since it unexports the Reason field;
users of the QuorumCallError must update their code to use the
errors.Is() function.

This also adds the gorums.Incomplete sentinal error that can be
used to check for incomplete quorum calls.