fix!: replace ConfigurationFromRaw() with Configuration.Clone()
Breaking change.

This replaces the ConfigurationFromRaw function with a new method
Clone on the Configuration struct. This method is does not appear
to be used very often; it is only used once in the hotstuff repo.
Hence, it shouldn't be breaking too many users. To be honest, I'm
not quite sure why we have this function; most of the time users
would create a configuration for their call types with the specific
quorum spec. The only need for the Clone(qspec) call would be to
replace a configuration's qspec with another qspec. I'm keeping
the function for now, but we may consider removing it if we don't
find a more useful need for it.