fix(channel): guard against nil pointer dereference
The gorumsStream can be set to nil by a goroutine runnning the reconnect
method after a connection is established. Furthermore, the receiver
goroutine can unlock the streamMutex and do some processing while this
is happening. As a result, the gorumsStream will become nil which will
panic when the RecvMsg() method is called on the gorumsStream. A guard
is thus implemented to prevent this scenario.