refactor: simplify verification by returning invalid signature error
This returns an error also for invalid signatures, avoiding the extra
check of the valid bool outside the VerifySignature method just to
generate an invalid signature error. Thus, we can simply return the
error directly. The returned error can be checked if it was an
InvalidSignatureErr if necessary (as is done in tests). This is not used
in non-test code yet.