refactor(snowflake): use consts for bit lengths; adds InvalidMachineID
This uses const values for bit lengths of timestamp, shard, machine ID,
and sequenceNum. Thus, the different max values and bit masks are
calculated based on the const bit lengths.

This also adds the InvalidMachineID function to allow initializing a
snowflake instance to an invalid machine ID that must be set afterwards.