refactor: consolidate connection management in RawNode
This removes the old dial() method, which was no longer blocking after
gRPC's move away from Dial to NewClient.

Note that the RawManager still blocks on AddNode, since RawNode.connect
calls channel.ensureStream, which does try to make a connection.

In the future, we may try to redesign the manager's blocking behavior to
be more aligned with gRPC's non-blocking approach.