feat: add NewResponseMessage and setError
The NewResponseMessage and setError is meant to replace WrapMessage.
The WrapMessage function is used by generated code and will be removed
in a future commit together updated code generation.

This change avoids setting the error in the outermost handler
immediately after calling the user provided implementation. Instead we
set the error just prior to sending the response message back to the
client.