refactor: replace defaultInterceptor with custom func()
The default interceptor approach had the same Handler interface as the
other interceptors, making it a bit incompatible with the go func()
usage in NodeStream(). This moves the "default interceptor" inside the
receive loop as a func().

This also makes use of the setError and NewResponseMessage and performs
various nil checks.