fix: make mock package more flexible and support wrapperspb
This replaces use of the mock.Register() test helper function with
a more flexible solution that supports wrapperspb types. The only
caveat is that this requires pre-registration via TestMain, but
is much more flexible, since we can add custom services, methods
and message types for our testing needs without having to define
them via proto files that needs to be generated.

Note: this adds mock.GetValue, which will be used in a future test.