refactor: response handling to use Result[T] type for type safety
The Result[T] type replaces the non-generic, unexported response type.
The reason to export the type is to make it easier to implement
interceptors that need to use this type. Note that Result[T] is currently
only used as Result[proto.Message] instances (in the current channel and
as called from the existing calltypes (like quorumcall, multicast etc).

In a future commit Result[T] will also be instansiated as Result[Resp],
or Result[any], which can be used at the interceptor level.