refactor: partially decouple node and channel
This adds a nodeOptions argument to the new newNode function and
takes care of creating the grpc client and channel with the correct
parameters as defined in the node options. This will help to better
decouple the Manager from the Node struct, avoiding hard to
understand connections between the two structures.