refactor: replace WithNoConnect with InsecureDialOptions in tests
This replaces the WithNoConnect manager option with the new
InsecureDialOptions test helper in tests, since creating a new
manager and subsequent Configurations will no longer block on
trying to connect to servers; connection dialing happens instead
during the actual invocation of a RPC call.

This also removes the WithNoConnect option so that external users
also cannot use it.