refactor: reduce coupling between channel and node
This moves more fields to the channel where they are actually used
to avoid the tight coupling between node and channel. It also makes
the initialization of the connection's outgoing metadata (passed in
via the ManagerOptions: WithPerNodeMetadata and WithMetadata) more
suitable for the current design. The commit also improves lifecycle
management via the context and cancel functions by more clearly
expressing these in close().