refactor: QuorumCallError to remove replies field and support Unwarp
This removes the replies field from the quorum call error, since
the replies count is not related to the error itself. We are mostly
interested in the number of errors, and the actual response should
provide guidance on how many replies were received and processed.

This also NodeErrors() to provide the number of nodes that cause
an error. It also provides Unwrap() that allows to inspect each
individual node error for a specific error type.