refactor: make interceptor design more flexible with next argument
This changes the intercepter design to this:

type QuorumInterceptor[Req, Resp msg] func(ctx *ClientCtx[Req, Resp], next ResponseSeq[Resp]) ResponseSeq[Resp]

Allowing more flexibility through chained invocations. This was
necessary to allow user-defined interceptors, which previously
was not possible to implement because ClientCtx was not exported
and also the responses iterator was also not exported previously.
This is now possible by exporting ClientCtx and augmenting the
QuorumInterceptor func type with the next argument. Thus, we avoid
reaching into the ClientCtx to get access to the "previous"
response iterator.