refactor: rename MessageID to message_seq_no in ordering.proto
This avoids the protobuf vs Go impedance mismatch for initialisms like
message_id (in protobuf) which is compiled to MessageId in Go, which is
not a recommended identifier in Go. To avoid this problem altogether, we
simply change the field to message_seq_no, which becomes a valid Go
identifier MessageSeqNo.